Moving the date util functions to the correct class
